caddyhttp: Refactor and export SanitizedPathJoin for use in fastcgi
